Remove need for extra space at end of line
The wiki code removes spaces when the next
line starts an HTML tag.  It was overeager
and removed the space if the next line
started with \W, which includes parentheses
and quotation marks.  I fixed it to look
for only '<' instead.  Trailing spaces are
only needed between <a> tags, now.

I also tried to be more consistent with
putting book titles in italics instead of
quotation marks.  And to put actual quotes
in quotation marks instead of italics.

Fixed a few spelling errors.