Fix non-text/html MIME files
If a line starts with `class` in the first 4KB of a text file, `file(1)`
assumes it's a C++ file.  I rearranged the review of "Animal Farm" so
the word `class` no longer starts a line.

`file(1)` also looks for the presence of certain HTML tags to identify
files as being HTML.  One of the patterns it looks for is `<a href=`.
By adding a `target` attribute between the tag name and the `href`
attribute, I can evade that rule.

`file(1)` also looks for `<table`, and I haven't found a clean way to
trick it into ignoring these tags, yet.

Inspired by reading
[an article on quality commits](https://fatbusinessman.com/2019/my-favourite-git-commit)
where the author of the commit used `find(1)` and `file(1)` to scan for
files with encodings that stood out from the rest.