Fix Java code samples
The double backslashes in the regex was breaking the article.  I
replaced the first one with its unicode entity `&#92;`.

I used [hilite.me](http://hilite.me/) with its `colorful` style.  I used
this style instead of the previous `bw` because I think it is closer to
the other code samples.

For some reason, the Java code samples needed `&nbsp;` entities on empty
lines, or else the markdown would treat indented lines as "code" with a
grey background.